--- app/src/main/java/info/anwesha/iitp/special/GameFragment.java
@@ -1,42 +1,0 @@
-package info.anwesha.iitp.special;
-
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import info.anwesha.iitp.R;
-
-public class GameFragment extends Fragment {
-
-    private WebView webView;
-
-    public GameFragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View rootview = inflater.inflate(R.layout.fragment_game, container, false);
-        webView = rootview.findViewById(R.id.game_view);
-        return rootview;
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        webView.setWebViewClient(new WebViewClient());
-        webView.loadUrl("https://5de6b07d668a1b74c0882d39--agitated-agnesi-547910.netlify.com/");
-
-    }
-}
