+++ app/src/main/res/layout/fragment_game.xml
@@ -1,13 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".special.GameFragment">
+    tools:context=".game.GameFragment">
 
     <WebView
-        android:id="@+id/game_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:id="@+id/gameWebView"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        />
 
-</FrameLayout>
+</LinearLayout>
