+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -33,8 +33,8 @@
         tools:layout="@layout/fragment_pronite" />
 
     <fragment
-        android:id="@+id/nav_game_view"
-        android:name="info.anwesha.iitp.special.GameFragment"
+        android:id="@+id/nav_game"
+        android:name="info.anwesha.iitp.game.GameFragment"
         android:label="Challenging Game"
         tools:layout="@layout/fragment_game" />
     <fragment
